backend/requirements.txt
fastapi==0.111.*
uvicorn[standard]
sqlalchemy==2.0.*
alembic
psycopg2-binary
pydantic[email]==2.*
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
httpx
